src/pignoletto/requirements.txt
Flask==2.0.2
Flask_Login==0.5.0
Flask_RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
GeoAlchemy2==0.10.2
pandas==1.3.5
PyJWT==2.3.0
requests==2.22.0
SQLAlchemy==1.4.29
Werkzeug==2.0.2
psycopg2
tqdm==4.62.3
